pyproject.toml
[tool.poetry]
name = "beamazed-task"
version = "0.1.0"
description = "test task for beamazedmedia"
authors = ["reid efe <reidefe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.3"
scikit-learn = "^1.6.0"
xgboost = "^2.1.3"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"
openai = "^1.59.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
